+++ core/virtualization.py
@@ -1,4 +1,4 @@
-import docker
+import docker, time
 from docker import errors as de
 
 from core.core_log import get_logger