+++ fetch_forecasts.py
@@ -7,6 +7,7 @@ from core.core_configuration import load_config, database_config, dwd_config, go
     hometemp_config
 from core.core_log import setup_logging, get_logger
 from core.database import DwDDataHandler, GoogleDataHandler, UlmDeHandler, WetterComHandler
+from core.usage_util import init_database
 from endpoint.fetcher import DWDFetcher, GoogleFetcher, UlmDeFetcher, WetterComFetcher
 
 # GLOBAL Variables
@@ -113,6 +114,7 @@ def ulmde_fetch_and_save():
 
 def main():
     log.info(f"------------------- Fetch DWD Measurements v{hometemp_config()['version']} -------------------")
+    init_database(UlmDeHandler, database_config(), 'ulmde_data')
     schedule.every(10).minutes.do(ulmde_fetch_and_save)
     schedule.every(10).minutes.do(dwd_fetch_and_save)
     schedule.every(10).minutes.do(google_fetch_and_save)